Rename templates and static files
