+++ sass/etna.scss
@@ -49,7 +49,7 @@ These are Etna specific components, created using BEM and following our guidelin
 @import 'includes/hierarchy-local-nav';
 @import 'includes/key-facts';
 @import 'includes/image-viewer-panel';
-@import 'includes/insights';
+@import 'includes/article';
 @import 'includes/record-access-options';
 @import 'includes/record-details';
 @import 'includes/taxonomy-tags';
