+++ sass/includes/_article.scss
@@ -1,4 +1,4 @@
-.insights-container {
+.article-container {
     display: grid;
     display:-ms-grid;
     grid-template-columns: repeat(4, 1fr);
