+++ scripts/src/article.js
@@ -4,21 +4,21 @@ import add_analytics_data_card_position from './modules/analytics/card_position'
 import audio_tracking from "./modules/analytics/audio_tracking"
 import video_tracking from "./modules/analytics/video_tracking"
 import add_unique_ids from "./modules/analytics/add_unique_ids";
-import mobile_tracking from "./modules/analytics/insights_tracking/mobile_tracking";
-import remove_mobile_tracking from "./modules/analytics/insights_tracking/remove_mobile_tracking";
-import link_list_tracking from "./modules/analytics/insights_tracking/link_list_tracking";
+import mobile_tracking from "./modules/analytics/article_tracking/mobile_tracking";
+import remove_mobile_tracking from "./modules/analytics/article_tracking/remove_mobile_tracking";
+import link_list_tracking from "./modules/analytics/article_tracking/link_list_tracking";
 
-import accordion_functionality from "./modules/insights/accordion_functionality/accordion_functionality";
-import add_event from "./modules/insights/add_event";
-import add_section_ids from "./modules/insights/accordion_functionality/add_section_ids";
-import apply_aria_roles from "./modules/insights/accordion_functionality/apply_aria_roles";
+import accordion_functionality from "./modules/articles/accordion_functionality/accordion_functionality";
+import add_event from "./modules/articles/add_event";
+import add_section_ids from "./modules/articles/accordion_functionality/add_section_ids";
+import apply_aria_roles from "./modules/articles/accordion_functionality/apply_aria_roles";
 import debounce from "./modules/debounce";
-import jumplinks_smooth_scroll from './modules/insights/navigation/jumplinks_smooth_scroll';
-import open_first_section from "./modules/insights/accordion_functionality/open_first_section";
-import remove_aria_roles from "./modules/insights/accordion_functionality/remove_aria_roles";
-import remove_event from "./modules/insights/remove_event";
-import set_active from "./modules/insights/navigation/set_active";
-import set_heading_positions from "./modules/insights/accordion_functionality/set_heading_positions";
+import jumplinks_smooth_scroll from './modules/articles/navigation/jumplinks_smooth_scroll';
+import open_first_section from "./modules/articles/accordion_functionality/open_first_section";
+import remove_aria_roles from "./modules/articles/accordion_functionality/remove_aria_roles";
+import remove_event from "./modules/articles/remove_event";
+import set_active from "./modules/articles/navigation/set_active";
+import set_heading_positions from "./modules/articles/accordion_functionality/set_heading_positions";
 import throttle from "./modules/throttle";
 
 document.addEventListener('DOMContentLoaded', () => {
@@ -43,14 +43,14 @@ window.addEventListener('load', () => {
     const remScreenSize = 48;
     let headingPositions;
 
-    // These booleans are used to detect if certain enhancements (e.g. click event listeners) have been applied in order to 
+    // These booleans are used to detect if certain enhancements (e.g. click event listeners) have been applied in order to
     // prevent events from being added more than once.
     let mobileEnhancementsApplied = false;
     let desktopEnhancementsApplied = false;
 
     // Ids are added to sections for ARIA purposes.
     add_section_ids($sectionHeadings, $sectionContents);
-    
+
     if($(window).width() / baseFontSize <= remScreenSize && !mobileEnhancementsApplied) {
         $sectionContents.hide();
         apply_aria_roles($sectionHeadings, $sectionContents);
@@ -77,7 +77,7 @@ window.addEventListener('load', () => {
         mobileEnhancementsApplied = true;
     }
     else if(!desktopEnhancementsApplied) {
-        // Add click and scroll listeners for smooth scrolling when using jumpinks to navigate the page and 
+        // Add click and scroll listeners for smooth scrolling when using jumpinks to navigate the page and
         // progress indication in the navigation on desktop.
         add_event($(window), "scroll", throttle(function() {
             set_active($sectionHeadings);
@@ -125,12 +125,12 @@ window.addEventListener('load', () => {
         else if ($(window).width() / baseFontSize > remScreenSize && !desktopEnhancementsApplied) {
             remove_aria_roles($sectionHeadings);
 
-            // Remove click and enter listeners because sections are fully expanded on desktop 
+            // Remove click and enter listeners because sections are fully expanded on desktop
             // i.e. no accordion functionality required.
             remove_event($sectionHeadings, "click");
             remove_event($sectionHeadings, "keypress");
 
-            // Add click and scroll listeners for smooth scrolling when using jumpinks to navigate the page and 
+            // Add click and scroll listeners for smooth scrolling when using jumpinks to navigate the page and
             // progress indication in the navigation when the screen size is increased.
             add_event($(window), "scroll", throttle(function() {
                 set_active($sectionHeadings);
@@ -142,7 +142,7 @@ window.addEventListener('load', () => {
             $sectionContents.show();
 
             remove_mobile_tracking();
-            
+
             desktopEnhancementsApplied = true;
             mobileEnhancementsApplied = false;
         }
