+++ scripts/src/modules/analytics/article_tracking/remove_mobile_tracking.js
@@ -1,5 +1,5 @@
 export default function remove_mobile_tracking() {
-    const insights_body = document.querySelector(".insights-container__main");
+    const article_body = document.querySelector(".article-container__main");
     const section_headings = document.querySelectorAll(".section-separator__heading");
 
     Array.prototype.forEach.call(section_headings, item => {
@@ -9,7 +9,7 @@ export default function remove_mobile_tracking() {
         item.removeAttribute("data-link");
     });
 
-    insights_body.removeEventListener("click", e => {
+    article_body.removeEventListener("click", e => {
         mobile_section_tracking(e);
     })
 }
