+++ scripts/src/modules/analytics/article_tracking/set_mobile_data_link_type.js
