+++ scripts/src/modules/analytics/article_tracking/set_mobile_tracking_attributes.js
