+++ scripts/src/modules/articles/accordion_functionality/accordion_functionality.js
