+++ scripts/src/modules/articles/accordion_functionality/add_section_ids.js
