+++ scripts/src/modules/articles/accordion_functionality/apply_aria_roles.js
