+++ scripts/src/modules/articles/accordion_functionality/open_first_section.js
