+++ scripts/src/modules/articles/accordion_functionality/remove_aria_roles.js
