+++ scripts/src/modules/articles/accordion_functionality/scroll_to_active_heading.js
