+++ scripts/src/modules/articles/accordion_functionality/set_heading_positions.js
