+++ scripts/src/modules/articles/accordion_functionality/slide_toggle.js
