+++ scripts/src/modules/articles/accordion_functionality/toggle_aria_expanded.js
