+++ scripts/src/modules/articles/navigation/jumplinks_smooth_scroll.js
