+++ scripts/src/modules/articles/remove_event.js
