+++ templates/articles/article_index_page.html
@@ -1,0 +1,80 @@
+{% extends 'base_page.html' %}
+
+{% load static %}
+{% load i18n %}
+{% load wagtailcore_tags %}
+{% load wagtailimages_tags %}
+{% load wagtailmetadata_tags %}
+{% block meta_tag %}
+  {% meta_tags %}
+{% endblock %}
+{% block content %}
+    {% include 'includes/generic-intro.html' %}
+    <div class="container">
+        <h2 class="sr-only">{% trans "Each collection insight takes an in-depth look into a group of the records we hold." %}</h2>
+    </div>
+
+    {% if page.featured_article %}
+    <div class="container">
+        <div class="featured-article">
+            {% if page.featured_article.hero_image %}
+                <a href="{% pageurl page.featured_article %}" class="featured-article__image-link" tabindex="-1" aria-hidden="true" data-component-name="Featured Story: {{ page.featured_article.title }}" data-link-type="Image" {% if page.featured_article.first_published_at.date > page.new_label_end_date.date %}data-label = "New"{%endif%}>
+                    <picture>
+                        {% image page.featured_article.teaser_image fill-500x400-c100 as hero %}
+                        <source media="(max-width: 768px)" srcset="{{ hero.url }}">
+                        {% image page.featured_article.teaser_image fill-500x500-c100 as hero %}
+                        <source media="(max-width: 991px)" srcset="{{ hero.url }}">
+                        {% image page.featured_article.teaser_image fill-1000x600-c100 as hero %}
+                        <source media="(max-width: 1200px)" srcset="{{ hero.url }}">
+
+                        {% if page.featured_article.hero_image_decorative %}
+                            <img src="{{ hero.url }}" alt="" />
+                        {% else %}
+                            <img src="{{ hero.url }}" alt="{{page.featured_article.hero_image_alt_text}}" />
+                        {% endif %}
+                    </picture>
+                </a>
+            {% endif %}
+
+            <div class="featured-article__description">
+                <h2 class="featured-article__heading">
+                    <small>{% trans "Spotlight on" %}</small>
+                    {{ page.featured_article.title }}
+                    {% if page.featured_article.first_published_at.date > page.new_label_end_date.date %}
+                        <span aria-describedby="article-desc" role="link">NEW</span>
+                    {% endif %}
+                </h2>
+
+                <p id ="article-desc" class="aria-desc">This is a link to a new story about {{ page.featured_article.title }}</p>
+
+                <p>{{ page.featured_article.sub_heading }}</p>
+                <a class="tna-button--dark" href="{% pageurl page.featured_article %}" data-component-name="Featured Story: {{ page.featured_article.title }}" data-link-type="Button" {% if page.featured_article.first_published_at.date > page.new_label_end_date.date %}data-label = "New"{%endif%}>Read</a>
+            </div>
+        </div>
+    </div>
+    {% endif %}
+
+    <div class="container">
+        {% include_block page.featured_pages %}
+    </div>
+
+    <div class="container mt-4" data-container-name="{{ page.title }}">
+        <h2>Discover all stories</h2>
+        <p>Browse and explore the human stories behind The National Archives’ collection.</p>
+        <ul class="card-group--list-style-none">
+
+            {% for article_page in article_pages %}
+
+                        {% with page=article_page.specific %}
+                         {% include 'includes/card-group-secondary-nav.html' with heading="Discover all stories" %}
+                        {% endwith %}
+
+            {% endfor %}
+        </ul>
+    </div>
+
+{% endblock %}
+
+{% block extra_js %}
+    <script src="{% static 'scripts/article_index_page.js' %}"></script>
+{% endblock %}
