+++ templates/articles/blocks/featured_record.html
