+++ templates/articles/blocks/images/blog-embed__image-container.html
